将表数量变为3

+++ src/main/java/com/example/myshardingjdbc/mapper/TImageMapper.java
@@ -21,7 +21,7 @@ public interface TImageMapper {
 
     List<TImage> selectByExample(TImageExample example);
 
-    TImage selectByPrimaryKey(Integer imageId);
+    TImage selectByPrimaryKey(Long imageId);
 
     int updateByPrimaryKeySelective(TImage record);
